Move tests out of stable module
In the future this will help test that stable and unstable expose the same API.
